Fix cli requests: Register our RequestHandlerInterface only for HTTP requests
This is happending because \TYPO3\CMS\Core\Core\Bootstrap::getInstance()->registerRequestHandlerImplementation()
is considered internal API and should not be used (from us). we used it unconditionally
and always added our Request Handler.

We need to make sure we don't do this for CLI requests as the interface for
the cli ReqestHandler receive a Symfony\Component\Console\Input\ArgvInput
instead of Psr\Http\Message\ServerRequestInterface.
As the types are declared in the respective RequestHandlerInterfaces
we can only implement one RequestHandlerInterfaces, not both.

The Exception was:

Uncaught TYPO3 Exception Argument 1 passed to Bnf\SlimTypo3\Http\SlimRequestHandler::canHandleRequest()
must be an instance of Psr\Http\Message\ServerRequestInterface, instance of
Symfony\Component\Console\Input\ArgvInput given, called in
[…]/vendor/typo3/cms/typo3/sysext/core/Classes/Core/Bootstrap.php on line 280
thrown in file […]/slim_typo3/Classes/Http/SlimRequestHandler.php in line 88