Introduce a new AppRegistry
Which is just a \SplStack implemented as Singleton,
which qualifies it as a global registry.
We use a stack as we want the executing of the registered
app' functions to be reverse.
As e.g. middleware's that are registered later
are called first (as they are stored on a stack as well).
By using the stack it's easier to wrap stuff from one extension
a in an extension b that depends on a.

The registry needs to be called non-static, means
GeneralUtility::makeInstance boilerplate in ext_localconf.php.
Sorry, but the App::register was ugly to test..

This is in preparation for Slim 4 where we'll
drop \Bnf\SlimTypo3\App, as the \Slim\App
class in Slim 4 will be generic enough
to not need extending (route-calculation
has been moved into a middleware).