Bundle bfr/typo3-middleware in the upcoming ter release
Also rework the build script from composer.json to Makefile.
This reduces the size of the composer.json, which is good, as that's
what we ship in composer/ter builds.

The Makefile is ignored in .gitattributes – so it's
fine to include the build there.
Sitenote: And of course good old unix make
is a much better build tool than composer scripts.
(We had them only for windows compatibility,
but that's really no focus here)