Support PSR-15 middlewares in slim
We build a shim bnf/slim3-psr15 to add support
for PSR-15 middlewares in slim3.

We use the concepts and one class of this packages.
As we still support the TYPO3 flavored class->method
description we still need to keep our own CallableResolver,
Maybe we make bnf/slim3-psr15 configurable at some point
to allow configuration of the callable description and the
instantiator.