Support handler instantiation using the system-wide DI container
Wrap our Pimple and the system wide symfony (with fallback to Extbase ObjectManager)
into a merging PSR-11 container which is provided to the Slim
middleware/route dispatchers.

Also gelegate the pimple container to return results from system container,
to allow service configuration to be mixed.

If the (symfony) system container is not available (which is the
case for TYPO3 v9), we fallback to the Extbase ObjectManager.