ext_emconf: Mark v8 compatible to avoid problems with TYPO3-Console upgrade:all
We used to mark v8 + v9 compatible in composer.json, but only
v9 comaptible in ext_emconf.php, as this extension does no longer
work in v8 classic mode (because of missing dependencies).

Now TYPO3-Console parses ext_emconf.php in upgrade:all, and throws an error:
  ./vendor/bin/typo3cms upgrade:all
  Incompatible extensions found, aborting.
  "slim_typo3" requires TYPO3 versions 9.5.0 - 9.5.99.
  It is not compatible with TYPO3 version "8.7.30"

Therefore we need to update that information to reflect the state that we
are still v8 compatible in composer mode.

This change will not be uploaded to TER.

Thanks to Sebastian Michaelsen (@smichaelsen) for pointing this out.